+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 - [2026-06-17T11:08:01.087Z] feat: create AppInterface container
 - [2026-06-17T12:54:34.230Z] refactor: abstract AppInterface state
 - [2026-06-17T13:00:05.882Z] feat: implement Tracker tab
+- [2026-06-17T18:30:42.128Z] feat: add dashboard components